% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bootcoefs.lmrob}
\alias{bootcoefs.lmrob}
\title{Bootstrap the coefficients from a single robust regression model}
\usage{
\method{bootcoefs}{lmrob}(object, R = 999, method = c("frb", "residuals",
  "cases"), ncpus = NULL, cl = NULL, ...)
}
\description{
Bootstrap the coefficients from a single robust regression model
}

