% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootcoefs-methods.R
\name{bootcoefs}
\alias{bootcoefs}
\alias{bootcoefs.complmrob}
\alias{bootcoefs.lmrob}
\title{Bootstrap the regression coefficients for a robust linear regression model}
\usage{
bootcoefs(
  object,
  R = 999,
  method = c("frb", "residuals", "cases"),
  ncpus = NULL,
  cl = NULL,
  ...
)

\method{bootcoefs}{complmrob}(
  object,
  R = 999,
  method = c("frb", "residuals", "cases"),
  ncpus = NULL,
  cl = NULL,
  ...
)

\method{bootcoefs}{lmrob}(
  object,
  R = 999,
  method = c("frb", "residuals", "cases"),
  ncpus = NULL,
  cl = NULL,
  ...
)
}
\arguments{
\item{object}{the model to bootstrap the coefficients from}

\item{R}{the number of bootstrap replicates.}

\item{method}{one of \code{"frb"} for fast and robust bootstrap, \code{"residuals"} to resample
the residuals or \code{"cases"} to resample the cases.}

\item{ncpus}{the number of CPUs to utilize for bootstrapping.}

\item{cl}{a snow or parallel cluster to use for bootstrapping.}

\item{...}{currently ignored.}
}
\value{
A list of type \code{bootcoefs} for which \code{\link[=print.bootcoefs]{print}},
     \code{\link{summary}} and \code{\link[=plot.bootcoefs]{plot}} methods are available
}
\description{
This function provides an easy interface and useful output to bootstrapping the regression
coefficients of robust linear regression models
}
\details{
If `object` is created by `complmrob` the default method is to use fast and robust bootstrap (FRB) as
described in the paper by M. Salibian-Barrera, et al (2008). The same default is used if `object` is an MM-estimate
created by `lmrob(..., method = 'SM')`.
The other options are to bootstrap the residuals or to bootstrap cases (observations), but the sampling
distribution of the estimates from these methods can be numerically unstable and take longer to compute.
If the `object` is a robust estimate created by `lmrob`, but not an MM-estimate, the default is to bootstrap
the residuals.
}
\section{Methods (by class)}{
\itemize{
\item \code{bootcoefs(complmrob)}: For robust linear regression models with compositional data

\item \code{bootcoefs(lmrob)}: For standard robust linear regression models

}}
\examples{
data <- data.frame(lifeExp = state.x77[, "Life Exp"], USArrests[ , -3])
mUSArr <- complmrob(lifeExp ~ ., data = data)
bc <- bootcoefs(mUSArr, R = 200) # the number of bootstrap replicates should
                                 # normally be higher!
summary(bc)
plot(bc) # for the model diagnostic plots
}
\references{
M. Salibian-Barrera, S. Aelst, and G. Willems. Fast and robust bootstrap. Statistical Methods and Applications, 17(1):41-71, 2008.
}
