% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bootStatCases}
\alias{bootStatCases}
\title{Function to calculate the coefficient(s) of the robust linear regression model
from a bootstrapped sample}
\usage{
bootStatCases(dat, inds, coefind, formula, maxTries = 4L)
}
\arguments{
\item{dat}{The original data set}

\item{inds}{The resampled indices}

\item{coefind}{The index of the coefficient to extract}

\item{formula}{the formula to fit the model}

\item{maxTries}{The maximum number of tries to increase the maxit control arguments for the S estimator}
}
\description{
Function to calculate the coefficient(s) of the robust linear regression model
from a bootstrapped sample
}

