% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bootStatFast}
\alias{bootStatFast}
\title{Function to calculate the bootstrapped coefficient of a robust linear regression model}
\usage{
bootStatFast(data, inds, control, coefind)
}
\arguments{
\item{data}{The original data frame}

\item{inds}{The resampled indices}

\item{control}{The control object as returned by \code{\link{bootStatFastControl}}.}

\item{coefind}{The index of the coefficient to extract}
}
\description{
Function to calculate the bootstrapped coefficient of a robust linear regression model
}
\references{
Salibian-Barrera M., Zamar R.: Bootstrapping robust estimates of regression, 2002
}

