% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bootStatFastControl}
\alias{bootStatFastControl}
\title{Function to extract the necessary data from the model and to calculate the correction
factors for the fast and robust bootstrap}
\usage{
bootStatFastControl(model)
}
\arguments{
\item{model}{The lmrob model}
}
\description{
Function to extract the necessary data from the model and to calculate the correction
factors for the fast and robust bootstrap
}
\references{
Salibian-Barrera M., Zamar R.: Bootstrapping robust estimates of regression, 2002
}

