% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bootStatResiduals}
\alias{bootStatResiduals}
\title{Function to calculate the coefficient(s) of the robust linear regression model
from a bootstrapped sample}
\usage{
bootStatResiduals(dat, inds, coefind, intercept = TRUE, maxTries = 4L)
}
\arguments{
\item{dat}{The original data set (with the columns fit, resid and the predictor variables)}

\item{inds}{The resampled indices}

\item{coefind}{The index of the coefficient to extract}

\item{intercept}{If the model includes an intercept term}

\item{maxTries}{The maximum number of tries to increase the maxit control arguments for the S estimator}
}
\description{
Function to calculate the coefficient(s) of the robust linear regression model
from a bootstrapped sample
}

