% Generated by roxygen2 (4.0.0): do not edit by hand
\name{complmrob}
\alias{complmrob}
\title{Function for robust linear regression using compositional predictors}
\usage{
complmrob(formula, data)
}
\arguments{
\item{formula}{The formula for the regression model}

\item{data}{The data.frame to use}
}
\value{
A list of type \code{complmrob} with fields
}
\description{
Function for robust linear regression using compositional predictors
}

