% Generated by roxygen2 (4.0.0): do not edit by hand
\name{confint.bclmrob}
\alias{confint.bclmrob}
\title{Extract the confidence intervals of the coefficients from the bootstrapped lmrob object}
\usage{
\method{confint}{bclmrob}(object, param, level = 0.95, type = c("bca",
  "perc", "norm", "basic", "stud"), ...)
}
\arguments{
\item{object}{the returned object from a call to bootcoefs}

\item{param}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{type}{the type of interval required (see the type argument of \code{\link{boot.ci}}).}

\item{...}{Currently ignored}
}
\description{
Extract the confidence intervals of the coefficients from the bootstrapped lmrob object
}

