% Generated by roxygen2 (4.0.1): do not edit by hand
\name{isomLRinv}
\alias{isomLRinv}
\title{Inverse Isometric log-ratio transformation for compositional data}
\usage{
isomLRinv(x, perc = TRUE)
}
\arguments{
\item{x}{a numeric vector of length \code{p-1} or a numeric matrix with \code{p-1} columns}

\item{perc}{should the result be a matrix with percentual shares (default \code{TRUE}).}
}
\value{
a numeric matrix with \code{p} columns with the transformed values. The values in the matrix
     are not on the original scale, but the percentual shares are equal
}
\description{
Projects the isometric log-ratio transformed (p-1) dimensional data back to the p-dimensional space
}

