% Generated by roxygen2 (4.0.0): do not edit by hand
\name{predictdf.complmrob.part}
\alias{predictdf.complmrob.part}
\title{This function is used by ggplot2 to predict the values for a \code{complmrob} model}
\usage{
predictdf.complmrob.part(model, xseq, se, level)
}
\arguments{
\item{model}{the complmrob.part model the prediction should be done for}

\item{xseq}{the sequence of x values to predict for}

\item{se}{should the confidence interval be returned as well}

\item{level}{the level of the confidence interval (if any)}
}
\description{
The function is exported solely because the ggplot2 method \code{predictdf} is not exported
and thus the generics system would not be working otherwise.
}

