% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.complmrob}
\alias{print.bootcoefs}
\alias{print.complmrob}
\title{Print the object}
\usage{
\method{print}{complmrob}(x, conf.level = 0.95, ...)

\method{print}{bootcoefs}(x, conf.level = 0.95, conf.type = "perc", ...)
}
\arguments{
\item{x}{the object to be printed.}

\item{conf.level}{the confidence level for the confidence interval.}

\item{conf.type}{the type of the printed confidence interval.}

\item{...}{ignored.}
}
\description{
Print information about the models returned by \code{\link{complmrob}} or \code{\link{bootcoefs}}.
For a detailed description see the help on \code{\link[=summary.complmrob]{summary}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{complmrob}: for robust linear regression models with compositional data

\item \code{bootcoefs}: for bootstrapped robust linear regression models with or without compositional data
}}
\seealso{
\code{\link{summary.complmrob}}
}

