% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-methods.R
\name{summary-methods}
\alias{summary-methods}
\alias{summary.complmrob}
\alias{summary.bccomplmrob}
\alias{summary.bclmrob}
\title{Get summary information}
\usage{
\method{summary}{complmrob}(object, conf.level = 0.95, ...)

\method{summary}{bccomplmrob}(object, conf.level = 0.95, conf.type = "perc", ...)

\method{summary}{bclmrob}(object, conf.level = 0.95, conf.type = "perc", ...)
}
\arguments{
\item{object}{the object for which the summary information should be returned.}

\item{conf.level}{the level of the returned confidence intervals.}

\item{...}{ignored.}

\item{conf.type}{the type of the returned confidence interval (see \code{\link[boot]{boot.ci}} for the
meaning of this parameter).}
}
\description{
List the estimates, standard errors, p-values and confidence intervals for the coefficients of
robust linear regression models with compositional data as returned by \code{\link{complmrob}} or
\code{\link{bootcoefs}}
}
