% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.complmrob}
\alias{summary.bccomplmrob}
\alias{summary.bclmrob}
\alias{summary.complmrob}
\title{Get summary information}
\usage{
\method{summary}{complmrob}(object, conf.level = 0.95, ...)

\method{summary}{bccomplmrob}(object, conf.level = 0.95, conf.type = "perc",
  ...)

\method{summary}{bclmrob}(object, conf.level = 0.95, conf.type = "perc",
  ...)
}
\arguments{
\item{object}{the object for which the summary information should be returned.}

\item{conf.level}{the level of the returned confidence intervals.}

\item{conf.type}{the type of the returned confidence interval (see \code{\link[boot]{boot.ci}} for the
meaning of this parameter).}

\item{...}{ignored.}
}
\description{
List the estimates, standard errors, p-values and confidence intervals for the coefficients of
robust linear regression models with compositional data as returned by \code{\link{complmrob}} or
\code{\link{bootcoefs}}
}
\section{Methods (by class)}{
\itemize{
\item \code{complmrob}: for robust linear regression models with compositional data

\item \code{bccomplmrob}: for bootstrapped robust linear regression models with compositional data

\item \code{bclmrob}: for bootstrapped robust linear regression models
}}

