% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bootcoefs.complmrob}
\alias{bootcoefs.complmrob}
\title{Bootstrap the coefficients from the compositional robust regression model}
\usage{
\method{bootcoefs}{complmrob}(object, R = 999, method = c("frb",
  "residuals", "cases"), ncpus = NULL, cl = NULL, ...)
}
\description{
Bootstrap the coefficients from the compositional robust regression model
}

