% Generated by roxygen2 (4.0.0): do not edit by hand
\name{complmrob}
\alias{complmrob}
\title{Function for robust linear regression using compositional predictors}
\usage{
complmrob(formula, data, boottype = c("fast", "residuals", "cases"),
  R = 100, ncpus = NULL, cl = NULL)
}
\arguments{
\item{formula}{The formula for the regression model}

\item{data}{The data.frame to use}

\item{boottype}{One of \code{fast}, \code{residuals} or \code{cases} (see Details for more information)}

\item{R}{The number of bootstrap replicates to generate to estimate the standard error and confidence
intervals. If NULL or missing, the theoretical standard error and confidence intervals are returned.}

\item{boot.params}{a list of parameters additionally passed to the boot function for bootstrapping
(see \code{\link{boot}} for available arguments). Especially interesting if parallel execution
is required.}
}
\value{
A list of type \code{complmrob} with fields
}
\description{
Function for robust linear regression using compositional predictors
}

