% Generated by roxygen2 (4.0.1): do not edit by hand
\name{format.perc}
\alias{format.perc}
\title{Simple function (just copied from the stats package) to format percentages}
\usage{
\method{format}{perc}(probs, digits)
}
\arguments{
\item{probs}{the percentages}

\item{digits}{the number of digits}
}
\description{
Simple function (just copied from the stats package) to format percentages
}

