% Generated by roxygen2 (4.0.1): do not edit by hand
\name{isomLR}
\alias{isomLR}
\alias{isomLRinv}
\title{Isometric log-ratio transformation for compositional data}
\usage{
isomLR(x, comp = 1)

isomLRinv(y, perc = TRUE)
}
\arguments{
\item{x}{a numeric vector of length \code{p} or a numeric matrix with \code{p} columns}

\item{comp}{the component to use as the first compositional part}

\item{y}{a numeric vector of length \code{D-1} or a numeric matrix with \code{D-1} columns}

\item{perc}{should the result be a matrix with percentage shares (default \code{TRUE}).}
}
\value{
a numeric matrix with \code{(p-1)} columns with the transformed values. The name of the first
    	column is the name of the first part (the other names are according to the order of the
		columns in the given matrix \code{x})

a numeric matrix with \code{D} columns with the transformed values. The values in the matrix
     are not on the original scale, but the percentage shares are equal
}
\description{
Projects the D-dimensional compositional data on the (D-1)-dimensional simplex isometrically by
transforming the values according to
\deqn{x_i = \sqrt{\frac{D - i}{D -i + 1}}  \log{ \frac{x_i}{ \left( \prod_{j=i+1}^{D} x_j \right)^{1/(D - i)} }}}{
 x_i = \sqrt((D - i)/(D - i + 1)) log( x_i / (\prod (j = i + 1 ... D) x_j)^(1/(D - i)) }

Projects the isometric log-ratio transformed (D-1) dimensional data back to the D-dimensional space
}
\section{Functions}{
\itemize{
\item \code{}: 
}}
\examples{
X <- as.matrix(USArrests[ , -3])
# Get the ilr with relative information of the 1st column to the other cols
ilrX1 <- isomLR(X)
# Get the ilr with relative information of the 2nd column to the other cols
irlX2 <- isomLR(X, 2)
isomLRinv(ilrX1)
}

