% Generated by roxygen2 (4.0.1): do not edit by hand
\name{isomLR}
\alias{isomLR}
\title{Isometric log-ratio transformation for compositional data}
\usage{
isomLR(x, comp = 1)
}
\arguments{
\item{x}{a numeric vector of length \code{p} or a numeric matrix with \code{p} columns}

\item{comp}{the component to use as the first compositional part}
}
\value{
a numeric matrix with \code{(p-1)} columns with the transformed values. The name of the first
    	column is the name of the first part (the other names are according to the order of the
		columns in the given matrix \code{x})
}
\description{
Projects the p-dimensional compositional data on the (p-1)-dimensional simplex isometrically
}

