% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.complmrob}
\alias{plot.complmrob}
\title{Diagnostic plots for the robust regression model with compositional covariats}
\usage{
\method{plot}{complmrob}(x, y = NULL, type = c("response", "model"),
  se = TRUE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{The complmrob object to plot}

\item{type}{one of \code{"response"} to plot the response or \code{"model"} to get the standard
lmrob model diagnostic plots.}

\item{se}{should the confidence interval be shown in the response plot}

\item{conf.level}{if the confidence interval is shown in the response plot, this parameter sets
     the level of the confidence interval}
}
\description{
Diagnostic plots for the robust regression model with compositional covariats
}

