% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.complmrob}
\alias{plot.complmrob}
\title{Diagnostic plots for the robust regression model with compositional covariats}
\usage{
plot.complmrob(x, type = c("response", "model", "bootstrap"), se = TRUE,
  conf.level = 0.95, ...)
}
\arguments{
\item{x}{The complmrob object to plot}

\item{type}{One of \code{response}, \code{model} or \code{bootstrap}}

\item{se}{Should the confidence interval be shown in the response plot}

\item{conf.level}{If the confidence interval is shown in the response plot, this parameter sets
     the level of the confidence interval}
}
\description{
Diagnostic plots for the robust regression model with compositional covariats
}

