% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{predictdf.complmrob.part}
\alias{predictdf.complmrob.part}
\title{Predict values for a \code{complmrob.part} object}
\usage{
predictdf.complmrob.part(model, xseq, se, level)
}
\arguments{
\item{model}{the complmrob.part model the prediction should be done for}

\item{xseq}{the sequence of x values to predict for}

\item{se}{should the confidence interval be returned as well}

\item{level}{the level of the confidence interval (if any)}
}
\description{
This function is used by ggplot2 to predict the values for a \code{complmrob} model and should
usally not be needed by the user.
}
\details{
The sole reason that this function is visible is because the \code{\link[ggplot2]{ggplot}}
function \code{predictdf} is not exported and thus this function could not be used for
\code{complmrob.part} objects if it was not exported.
}

