% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print-methods}
\alias{print-methods}
\alias{print.bootcoefs}
\alias{print.complmrob}
\title{Print the object}
\usage{
\method{print}{complmrob}(x, conf.level = 0.95, ...)

\method{print}{bootcoefs}(x, conf.level = 0.95, conf.type = "perc", ...)
}
\arguments{
\item{x}{the object to be printed.}

\item{conf.level}{the confidence level for the confidence interval.}

\item{conf.type}{the type of the printed confidence interval.}

\item{...}{ignored.}
}
\description{
Print information about the models returned by \code{\link{complmrob}} or \code{\link{bootcoefs}}.
For a detailed description see the help on \code{\link[=summary.complmrob]{summary}}.
}
\seealso{
\code{\link{summary-methods}}
}

