% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.summary.complmrob}
\alias{print.summary.complmrob}
\title{Print the summary information}
\usage{
\method{print}{summary.complmrob}(x, digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{the summary object.}

\item{digits}{the number of digits for the reported figures}

\item{signif.stars}{should stars be displayed to show the significance of certain figures}

\item{...}{further arguments currently not used}
}
\description{
Print the summary information
}

